Separate types, Side and Order
