+++ cpp/lib/BUILD
@@ -2,7 +2,22 @@ package(default_visibility = ["//visibility:public"])
 
 cc_library(
     name = "map-order-book",
-    hdrs = ["MapOrderBook.h"]
+    hdrs = ["MapOrderBook.h"],
+    deps = ["order"]
+)
+
+cc_library(
+    name = "order",
+    hdrs = ["Order.h"],
+    deps = [
+        "side",
+        "types"
+    ]
+)
+
+cc_library(
+    name = "side",
+    hdrs = ["Side.h"]
 )
 
 cc_library(
@@ -10,3 +25,7 @@ cc_library(
     hdrs = ["SPSCQueue.h"]
 )
 
+cc_library(
+    name = "types",
+    hdrs = ["types.h"]
+)
