+++ cpp/lib/MapOrderBook.h
@@ -4,6 +4,8 @@
 // --------------
 // Define an order book using maps for the bids and asks.
 
+#include "lib/Order.h"
+
 #include <iostream>
 #include <iterator>
 #include <list>
@@ -11,23 +13,6 @@
 #include <numeric>
 #include <unordered_map>
 
-using OrderIdT = uint32_t;
-using SizeT = uint16_t;
-using PriceT = double;
-
-enum class Side
-{
-    BID,
-    ASK
-};
-
-struct Order
-{
-    OrderIdT oid;
-    Side side;
-    PriceT price;
-    SizeT size;
-};
 
 class ListPriceLevel
 {
