+++ cpp/lib/Order.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "lib/Side.h"
+#include "lib/types.h"
+
+struct Order
+{
+    OrderIdT oid;
+    Side side;
+    PriceT price;
+    SizeT size;
+};
