+++ cpp/lib/Side.h
@@ -1,0 +1,7 @@
+#pragma once
+
+enum class Side
+{
+    BID,
+    ASK
+};
