+++ cpp/lib/types.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <cstdint>
+
+using OrderIdT = uint32_t;
+using SizeT = uint16_t;
+using PriceT = double;
